refactor(calendar): extract intervals as utilities
